refactor: reorganize modules into new subdirectories
