+++ py_modules/unifideck/accounts/__init__.py
@@ -1,0 +1,1 @@
+from .account_manager import AccountManager
