+++ py_modules/unifideck/cdp/__init__.py
@@ -1,0 +1,2 @@
+from .cdp_inject import get_cdp_client, shutdown_cdp_client
+from .cdp_utils import create_cef_debugging_flag
