+++ py_modules/unifideck/cloud/__init__.py
@@ -1,0 +1,1 @@
+from .cloud_save import CloudSaveManager
