+++ py_modules/unifideck/steam/steamgriddb.py
@@ -13,7 +13,7 @@ from typing import Optional, List, Dict, Any
 from pathlib import Path
 
 # Import Steam user detection utility
-from py_modules.unifideck.steam_utils import get_logged_in_steam_user
+from py_modules.unifideck.steam.steam_utils import get_logged_in_steam_user
 
 logger = logging.getLogger(__name__)
